requirements.txt
flask >= 1.1
gunicorn>=23.0
python-dotenv
pymysql